% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooni.R
\name{ooni_incidents}
\alias{ooni_incidents}
\title{Search OONI censorship incidents}
\usage{
ooni_incidents(limit = 20)
}
\arguments{
\item{limit}{Max results (default 20)}
}
\value{
tibble: id, title, text, start_time, create_time,
update_time, event_type
}
\description{
Retrieve reported internet censorship incidents from the OONI database.
}
